refactor: file repackaging
