+++ app/src/main/java/com/sopt/now/main/ResponseUserInfoDto.kt
@@ -1,4 +1,4 @@
-package com.sopt.now.data
+package com.sopt.now.main
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
