+++ app/src/main/java/com/sopt/now/service/AuthService.kt
@@ -1,10 +1,10 @@
 package com.sopt.now.service
 
-import com.sopt.now.data.RequestSignInDto
-import com.sopt.now.data.RequestSignUpDto
-import com.sopt.now.data.ResponseSignInDto
-import com.sopt.now.data.ResponseSignUpDto
-import com.sopt.now.data.ResponseUserInfoDto
+import com.sopt.now.main.ResponseUserInfoDto
+import com.sopt.now.signIn.RequestSignInDto
+import com.sopt.now.signUp.RequestSignUpDto
+import com.sopt.now.signIn.ResponseSignInDto
+import com.sopt.now.signUp.ResponseSignUpDto
 import retrofit2.Call
 import retrofit2.http.Body
 import retrofit2.http.GET
