+++ app/src/main/java/com/sopt/now/signIn/RequestSignInDto.kt
@@ -1,16 +1,12 @@
-package com.sopt.now.data
+package com.sopt.now.signIn
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignUpDto(
+data class RequestSignInDto(
     @SerialName("authenticationId")
     val authenticationId: String,
     @SerialName("password")
     val password: String,
-    @SerialName("nickname")
-    val nickname: String,
-    @SerialName("phone")
-    val phone: String,
 )
