+++ app/src/main/java/com/sopt/now/signUp/RequestSignUpDto.kt
@@ -1,12 +1,16 @@
-package com.sopt.now.data
+package com.sopt.now.signUp
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class RequestSignInDto(
+data class RequestSignUpDto(
     @SerialName("authenticationId")
     val authenticationId: String,
     @SerialName("password")
     val password: String,
+    @SerialName("nickname")
+    val nickname: String,
+    @SerialName("phone")
+    val phone: String,
 )
