+++ app/src/main/java/com/sopt/now/signUp/ResponseSignUpDto.kt
@@ -1,10 +1,10 @@
-package com.sopt.now.data
+package com.sopt.now.signUp
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class ResponseSignInDto(
+data class ResponseSignUpDto(
     @SerialName("code")
     val status: Int,
     @SerialName("message")
