+++ app/src/main/java/com/sopt/now/signUp/SignUpViewModel.kt
@@ -1,13 +1,10 @@
-package com.sopt.now.viewModel
+package com.sopt.now.signUp
 
 import android.util.Log
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.sopt.now.ServicePool
-import com.sopt.now.data.RequestSignUpDto
-import com.sopt.now.data.ResponseSignUpDto
-import com.sopt.now.state.SignUpState
+import com.sopt.now.service.ServicePool
 import retrofit2.Call
 import retrofit2.Callback
 import retrofit2.Response
